Users can add any type of text(String, int, double). Design with ListTile (with title and subtitle).
When Users long press on any list tile then show an Alertdialog with edit and delete button.

After that, Users can update the text with the edit button and delete the text by delete button.

(When pressing the edit button, open a bottom sheet with title and description text of selected list tile item)